rename: components 폴더에 하위 폴더 생성 및 파일 이동
