+++ src/components/style/CaptionStyle.tsx
