+++ src/components/style/DelimiterStyle.tsx
