+++ src/pages/LastPage/index.tsx
@@ -1,10 +1,10 @@
 import { ChangeEvent, KeyboardEventHandler, useContext, useState } from "react";
 import styled, { css } from "styled-components";
-import { NextButton } from "components/ButtonStyle";
+import { NextButton } from "components/style/ButtonStyle";
 import CardPreview from "components/CardPreview";
 import Input, { CardNickname } from "components/Input";
 import LengthLimit from "components/LengthLimit";
-import { CardInfoContext } from "components/CardInfoProvider";
+import { CardInfoContext } from "components/provider/CardInfoProvider";
 import GotLost from "pages/GotLost";
 import useSetCardInfo from "hooks/useSetCardInfo";
 import { changeInvalidValueToBlank } from "utils/inputValidator";
