+++ src/pages/RegisterPage/FormInputs/ExpirationDateInput.tsx
@@ -2,10 +2,10 @@ import { useFocus } from "hooks/useFocus";
 import { ChangeEvent, Fragment, useContext } from "react";
 import { changeInvalidValueToBlank } from "utils/inputValidator";
 import Input, { CommonInputStyle } from "components/Input";
-import { Slash } from "components/DelimiterStyle";
-import { DateCaption } from "components/CaptionStyle";
-import { DateInputBox } from "components/InputBoxStyle";
-import { CardInfoContext } from "components/CardInfoProvider";
+import { Slash } from "components/style/DelimiterStyle";
+import { DateCaption } from "components/style/CaptionStyle";
+import { DateInputBox } from "components/style/InputBoxStyle";
+import { CardInfoContext } from "components/provider/CardInfoProvider";
 import { isInvalidDate } from "validation";
 import { ExpirationDate } from "types";
 import { DATE_INPUT, LIMIT_LENGTH, VALID_INPUT } from "constants/limit";
